requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
imagecodecs==2025.3.30
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy_loader==0.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pywin32==311
pyzmq==27.0.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.16.1
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
threadpoolctl==3.6.0
tifffile==2025.6.11
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
wcwidth==0.2.13
